fixed critical bug that didnt let the program terminate when encoding files larger than 2gb. changed index from int to long to do this.

+++ src/de/christian/f/brinkmann/ibus/Encoder.java
@@ -36,7 +36,7 @@ public class Encoder {
 			if (source.length() > getMaxDataSize()) {
 				try {
 					FileInputStream fin = new FileInputStream(source);
-					for (int index = 0; index < source.length(); index += getMaxDataSize()) {
+					for (long index = 0; index < source.length(); index += getMaxDataSize()) {
 						byte[] fileContent;
 						int overheadBytes;
 						int paddingBytes;
